+++ src/bfg_api/urls.py
@@ -16,6 +16,7 @@ urlpatterns = [
 
     # user login/out
     path('user-login/<str:params>/', views.UserLogin.as_view()),
+    path('user-seat/<str:params>/', views.UserSeat.as_view()),
     path('user-logout/<str:params>/', views.UserLogout.as_view()),
 
     # REST api